stack constraint refactoring (WIP commit)
